chore: creating new form for observation level
